Add MultiBsplineEvalHelper
