+++ src/spline2/MultiBsplineVGLH_OMPoffload.hpp
@@ -25,6 +25,7 @@
 
 #include "OpenMP/OMPstd.hpp"
 #include "spline2/MultiBsplineData.hpp"
+#include <spline2/MultiBsplineEvalHelper.hpp>
 
 namespace spline2offload
 {
